# server/docs/swagger.yaml
definitions:
  types.GetResultHandler:
    properties:
      result:
        type: string
    type: object
  types.GetSessionIdHandler:
    properties:
      token:
        type: string
    type: object
  types.GetStatusHandler:
    properties:
      status:
        type: string
    type: object
  types.GetTaskIdHandler:
    properties:
      task_id:
        type: string
    type: object
  types.User:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: http server
  title: My API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetSessionIdHandler'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Post a Login
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - text/plain
      responses:
        "200":
          description: Пользователь {login} зарегистрирован.
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Post a Register
      tags:
      - user
  /result/{task_id}:
    get:
      consumes:
      - application/json
      description: Get a tasks result by its id
      parameters:
      - description: Task Id
        in: path
        name: task_id
        required: true
        type: string
      - description: Bearer {auth_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetResultHandler'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorize
          schema:
            type: string
        "404":
          description: Task id not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a Result
      tags:
      - result
  /status/{task_id}:
    get:
      consumes:
      - application/json
      description: Get a tasks status by its id
      parameters:
      - description: Task Id
        in: path
        name: task_id
        required: true
        type: string
      - description: Bearer {auth_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetStatusHandler'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorize
          schema:
            type: string
        "404":
          description: Task id not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a Status
      tags:
      - status
  /task:
    post:
      consumes:
      - application/json
      description: make task and get tsk id
      parameters:
      - description: Bearer {auth_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetTaskIdHandler'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorize
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Post a Task
      tags:
      - task
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"